DESRESFF#424
in cases where the first atom in the system is a carbon, the
atomtype from iviparr would be C0. This conflicts with C0
atomtype in the amber forcefields, which causes problems when
merging the iviparr forcefield into a standard amber ff.
This fixes the problem by prefixing all iviparr generated atom types
with 'iv_'